refactor: split reporting.py into handlers/schemas/rendering package
Wave A piece 3 of 3 — split the 2015-line goldfive/reporting.py into a
goldfive/reporting/ package. The original file conflated three concerns:
the agent-facing contract (handlers), the over-the-wire tool schemas
(schemas), and LLM-visible response payload construction (rendering).

New layout:

* goldfive/reporting/handlers.py — ReportingToolSpec, the _handle_*
  async functions, BUILTIN_REPORTING_TOOLS, the drift-opt-in selector,
  declaration-key constants, _validate_required, the
  _classify_and_route_pin bridge that turns pin-freshness classifications
  into routing/refusal decisions, and the approval-handler orchestration
  (kept here per the team-lead-approved scope; deeper executor split is
  a separate concern).
* goldfive/reporting/schemas.py — JSON-schema parameter blocks
  (_SCHEMA_TASK_STARTED, _SCHEMA_TASK_PROGRESS, _SCHEMA_TASK_COMPLETED,
  _SCHEMA_TASK_FAILED, _SCHEMA_TASK_BLOCKED, _SCHEMA_NEW_WORK_DISCOVERED,
  _SCHEMA_PLAN_DIVERGENCE, _SCHEMA_AWAITING_APPROVAL,
  _SCHEMA_DECLARE_TASK_SKIPPED, _SCHEMA_DECLARE_TASK_NOT_NEEDED) plus
  the _object_schema helper. These go verbatim over the wire to LLMs as
  tool input schemas; the constants are byte-identical to the pre-split
  values.
* goldfive/reporting/rendering.py — payload construction for what the
  LLM sees on the response side: _build_plan_state and its helpers
  (_next_pending_with_completed_predecessors, _bare_agent_name), the F1
  _directive_ack, the idempotent / invalid_transition / refused
  response shapes, and the missing-arg rejection bodies. _build_plan_state
  output is byte-identical (the directive-ack contract is load-bearing
  for coordinator loop-prevention).
* goldfive/reporting/_internal.py — shared package-private utilities:
  argument coercion (_str/_float/_bool/_int), pin/plan resolution
  (_find_task_in_session, _resolve_task_id*, _resolve_effective_task_id,
  _reroute_if_superseded), the pin-freshness classifier
  (_classify_pin_freshness, _read_pin_revision, _read_plan_revision,
  _supersession_successor), _classify_transition, _rotate_after_terminal,
  _await_plan_stable, and the sink emitters
  (_emit_task_transition_refused, _emit_task_declaration_received,
  _emit_approval_requested). Plus the _TOOL_TARGET_STATUS /
  _TOOL_VALID_SOURCES / _TERMINAL_STATUSES tables and the _ACK sentinel.

The package __init__.py re-exports every previously-importable name —
both the public surface (BUILTIN_REPORTING_TOOLS, ReportingToolSpec,
REPORTING_TOOL_NAMES, DECLARATION_KIND_*, DECLARATIONS_KEY,
DRIFT_SELF_REPORTING_TOOL_NAMES, DRIFT_SELF_REPORTING_TOOLS,
LIFECYCLE_REPORTING_TOOLS, ReportingHandler, select_reporting_tools)
and the private symbols tests and docs reach into
(_handle_task_started, _emit_task_transition_refused,
_resolve_effective_task_id, _build_plan_state, _validate_required, …).
External callers see no API change.

No behaviour change. Handler signatures, schema dict contents, and
_build_plan_state output are preserved exactly.

Test plan:
- Full suite: 1988 passed, 126 skipped (no new failures).
- Reporting-specific subset (test_reporting, test_reporting_declarations,
  test_reporting_tool_required_validation, test_reporting_idempotency,
  test_task_id_wiring, test_task_binding_follows_revision,
  test_approval_flow, test_cancellation_stash_audit, test_state_rotation,
  test_tier1_loop_prevention, test_pin_versioning) all green.
- _state_audit_enabled autouse fixture in tests/conftest.py continues
  to pass (audited stash region in _handle_task_started is intact;
  test_cancellation_stash_audit C5/C6 still green).
